Update dependencies and add new benchmark kind (#9)
This commit updates several dependencies in the `Cargo.lock` file,
including
`derive_more`, `cc`, `clap`, `clap_derive`, `uuid`, and `winnow`. The
`iggy-bench-report` dependency in `Cargo.toml` is also updated to track
the
`master` branch. Additionally, a new benchmark kind,
`EndToEndProducingConsumerGroup`, is added to the sidebar and benchmark
selectors in the frontend. This change enhances the functionality by
allowing
users to select and work with the new benchmark kind, improving the
overall
usability of the application.