Update benchmark version and enhance UI with benchmark counts (#11)
This commit updates the version of the `iggy-bench-dashboard` packages
from 0.2.5 to 0.2.6 in both `Cargo.lock` and `Cargo.toml`. Additionally,
it enhances the frontend UI by displaying the count of different
benchmark types in the sidebar and benchmark kind selector. This
provides users with a quick overview of the number of benchmarks
available for each category, improving the usability and information
accessibility of the dashboard.